// frontend/src/types.ts
export enum RegisterButtonColor {
  SeaGreen = "SeaGreen",
  PastelRed = "PastelRed"
};

export enum RegisterButtonSize {
  Large = "Large",
  Small = "Small"
}

export type IconProps = {
  id?: string;
  className?: string;
};

export enum PillNames {
  Listen = "Listen",
  Copy = "Copy",
  Paste = "Paste",
  Play = "Play"
}